Prime Minister Ehsaas Program
Prime Minister Imran Khan has started Ehsaas Program to support needy and poor people in Pakistan. The 
government started sending 12000 cash to those who registered themselves in ehsaas program in Pakistan.  
Ehsaas is about the creation of a ‘welfare state’ by countering elite capture and leveraging 21st century 
tools—such as using data and technology to create precision safety nets; promoting financial inclusion and 
access to digital services; supporting the economic empowerment of women; focusing on the central role of 
human capital formation for poverty eradication, economic growth and sustainable development; and 
overcoming financial barriers to accessing health and post-secondary education. The program is for the 
extreme poor, orphans, widows, the homeless, the disabled, those who risk medical impoverishment, for the 
jobless, for poor farmers, for laborers, for the sick and undernourished; for students from low-income 
backgrounds and for poor women and elderly citizens.  This plan is also about lifting lagging areas where 
poverty is higher.